fix to minimum support iOS9.0 ,and remove repeat extension file
