+++ Sources/Operator+.swift
@@ -37,3 +37,34 @@ infix operator |>: ForwardPipe
 public func |> <T, U>(lhs: T, rhs: (T) throws -> U) rethrows -> U {
   return try rhs(lhs)
 }
+
+// https://github.com/apple/swift-evolution/blob/master/proposals/0121-remove-optional-comparison-operators.md
+// FIXME: comparison operators with optionals were removed from the Swift Standard Libary.
+// Consider refactoring the code to use the non-optional operators.
+public func < <T: Comparable>(lhs: T?, rhs: T?) -> Bool {
+    switch (lhs, rhs) {
+    case let (l?, r?): return l < r
+    case (nil, _?): return true
+    default: return false }
+}
+
+public func > <T: Comparable>(lhs: T?, rhs: T?) -> Bool {
+    switch (lhs, rhs) {
+    case let (l?, r?):
+        return l > r
+    default:
+        return rhs < lhs
+    }
+}
+
+
+public func == <T: Equatable>(lhs: T?, rhs: T?) -> Bool {
+    switch (lhs, rhs) {
+    case let (l?, r?):
+        return l == r
+    case (nil, nil):
+        return true
+    default:
+        return false
+    }
+}
